refactor: ♻️ ディレクトリ構成を変更
src/componentsにelementsとlayoutsのサブディレクトリを作成。今後、共通コンポーネントはelementsに格納するように変更。
