+++ src/components/layouts/index.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import type { FC, ReactNode } from 'react';
-import Footer from '../Footer';
-import NavBar from '../NavBar';
+import Footer from './components/Footer';
+import NavBar from './components/NavBar';
 
 type Props = {
   pageTitle: string;
