+++ src/pages/auth/is-new-user.tsx
@@ -1,11 +1,11 @@
 import type { NextPage } from 'next';
-import Layout from '../../components/Layout';
+import Layout from '../../components/layouts';
 import IsNewUser from '../../features/is-new-user';
 
-const Home: NextPage = () => (
+const IsNewUserPage: NextPage = () => (
   <Layout pageTitle="Is new user">
     <IsNewUser />
   </Layout>
 );
 
-export default Home;
+export default IsNewUserPage;
