+++ src/pages/index.tsx
@@ -1,10 +1,10 @@
 import type { NextPage } from 'next';
-import Layout from '../components/Layout';
+import Layout from '../components/layouts';
 
-const Home: NextPage = () => (
+const HomePage: NextPage = () => (
   <Layout pageTitle="Programmist">
     <div className="my-80 font-serif text-5xl text-center underline">Hello, Programmist World!</div>
   </Layout>
 );
 
-export default Home;
+export default HomePage;
